requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
playwright>=1.53.0
openai>=1.35.0
pywebview>=5.1
redis>=5.0.0
gunicorn>=22.0.0
python-dotenv>=1.0.1
